# docs/swagger.yaml
openapi: 3.0.3
info:
  title: 꾹(GGUK) API
  version: 0.1.0-draft
  description: |
    꾹 MVP API 초안. `docs/prd.md` 8장(API 설계 초안)을 OpenAPI로 옮긴 것으로,
    요청/응답 스키마·에러 코드 값은 각 피쳐 PR에서 확정한다.

    - 성공 응답은 공통 인터셉터가 `{ "data": ... }`로 감싼다. 반환할 데이터가 없는 성공은
      본문 없는 204 대신 `{ "data": { "ok": true } }`를 반환한다.
    - 에러 응답은 공통 예외 필터의 `{ "errorCode", "message" }` 포맷을 따른다.
    - 목록 API 페이지네이션은 **offset 기반**으로 통일한다: `page`(0부터 시작)·`pageSize` 쿼리 파라미터,
      응답에 `pagination { pageSize, page, hasNext }` 객체 포함. (지도용 핀 전체 조회만 예외)
    - 인증은 Firebase Authentication 익명 인증이다. 클라이언트가 받은 ID 토큰을
      `Authorization: Bearer <idToken>`으로 보낸다. 토큰 없이 호출할 수 있는 API는
      `GET /api/v1/invitations/{code}` 뿐이며, `POST /api/v1/users`는 **토큰은 필요하지만
      등록된 유저는 아직 없어도 되는** 예외다.
    - 제목에 **[TBD]** 가 붙은 API는 기획 변경이 진행 중이라 계약이 잠정이다.
servers:
  - url: https://api.gguk.org
    description: production
  - url: http://localhost:3000
    description: local
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/rooms:
    post:
      tags: [Room]
      summary: 공동방 생성
      description: 생성자가 방장(`owner_id`)이 된다. 초대 코드는 `invitations`가 별도로 관리한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, color]
              properties:
                name:
                  $ref: "#/components/schemas/RoomName"
                description:
                  type: string
                  maxLength: 30 # grapheme 단위(클라이언트 카운터와 동일)
                  nullable: true
                color:
                  $ref: "#/components/schemas/ColorKey"
      responses:
        "201":
          description: 생성된 방
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [Room]
      summary: 내가 속한 방 목록
      description: 나간 방(멤버십 soft delete)은 제외한다. 공동방 생성 유도 바텀시트용 보유 이력 판별에도 쓰인다.
      parameters:
        - name: showHasPlaceId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: 지정 시 각 방에 해당 장소(placeId)가 이미 저장되어 있는지(`hasPlace`) 및 매칭 핀 ID(`matchedPinId`)를 함께 반환 — "다른 방에 공유" 화면용
        - name: showUsers
          in: query
          required: false
          schema:
            type: string
            enum: ["true", "false"]
          description: '`true`면 각 방의 멤버 목록(`users`)을 함께 반환. 정확히 `true`/`false`만 허용한다'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoomSummary"
        "400":
          description: showHasPlaceId가 UUID가 아니거나 showUsers가 `true`/`false`가 아님 (`VALIDATION_ERROR`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/rooms/{roomId}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      tags: [Room]
      summary: 방 상세 조회
      description: |
        방 정보 + 핀 수·멤버 수 메타. 요청 유저의 멤버십을 검증한다.
        핀 목록은 포함하지 않는다 — `GET /api/v1/pins?roomId=`로 별도 조회한다.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/RoomDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Room]
      summary: 방 편집
      description: |
        이름·설명·색상 수정. **방장만** 가능.
        세 필드 모두 선택이지만 **최소 하나는 있어야 한다** — 빈 본문은 400 `VALIDATION_ERROR`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:
                  $ref: "#/components/schemas/RoomName"
                description:
                  type: string
                  maxLength: 30 # grapheme 단위(클라이언트 카운터와 동일)
                  nullable: true
                color:
                  $ref: "#/components/schemas/ColorKey"
      responses:
        "200":
          description: 수정된 방
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Room"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/rooms/{roomId}/owner:
    put:
      tags: [Room]
      summary: 방장 위임
      description: '**방장만** 가능. 위임 대상은 해당 방의 활성 멤버여야 한다. 방장이 방을 나가려면 이 API로 먼저 위임을 완료해야 한다.'
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nextOwnerId]
              properties:
                nextOwnerId:
                  type: string
                  format: uuid
      responses:
        "200":
          description: 위임 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/rooms/{roomId}/members/me:
    delete:
      tags: [Room]
      summary: 방 나가기
      description: |
        멤버십 soft delete. 파라미터 없이 요청하며, 서버가 요청 유저의 상태를 검증해 분기한다.
        - 일반 멤버: 나가기 처리.
        - **방장 + 다른 멤버 존재: 에러로 거절** — 먼저 방장 위임(`PUT /rooms/{roomId}/owner`)을 완료해야 한다. (errorCode 예: `OWNER_TRANSFER_REQUIRED`)
        - **방장 + 마지막 멤버: 나가기를 허용하고 방을 삭제(soft delete) 처리한다.** (PR 리뷰 확정)
        - 개인방은 나가기 대상이 아니다.
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: 나가기 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: 방장이 위임 없이 나가기를 요청함 — 위임 후 재시도 필요
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/rooms/{roomId}/members:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [Invitation]
      summary: 방 합류
      description: |
        body의 `inviteCode`로 `invitations`에서 활성 초대를 찾고, 그 초대의 방이 path의 `{roomId}`와
        같은지 검증한다. 불일치·무효 코드는 멤버 추가 전에 거절한다.
        **개인방의 코드는 거절한다.** 이미 멤버면 오류 대신 멱등 응답을 준다.
        나갔던 방에는 같은 코드로 다시 합류할 수 있다(이전 멤버십 행은 `deleted_at`을 유지).

        합류 완료 시 기존 멤버에게 `ROOM_MEMBER_JOINED`, 합류한 본인에게 `ROOM_JOINED_SELF`
        알림을 기록하고 푸시를 발송한다. 알림 발송이 실패해도 합류 자체는 성공으로 응답한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inviteCode]
              properties:
                inviteCode:
                  $ref: "#/components/schemas/InvitationCode"
      responses:
        "200":
          description: 합류 완료(또는 이미 멤버 — 멱등)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [Room]
      summary: 방 멤버 목록
      description: 방장 위임 대상 선택 모달용. 요청 유저의 멤버십을 검증한다.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoomMember"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/pins:
    get:
      tags: [Pin]
      summary: 핀 목록 조회
      description: |
        `roomId`의 핀 목록(좌표 포함). 요청 유저의 해당 방 멤버십을 검증한다.
        **`page`/`pageSize` 미지정 시 전체를 반환한다** — 지도 전체 보기는 페이지네이션 구분이 애매해 전체 조회를 보장한다(성능 저하 확인 시 리팩토링, PR 리뷰 확정). 리스트 뷰 등에서는 offset 기반 페이지네이션한다.
        **목록 정렬 5종** (`sort`, 기본 `all`) — 홈 카드 덱의 정렬 칩과 별개다.
        - `all` 전체 / `ggukPick` 저장일 오래된 순 상위 30% / `latest` 저장일 최신순
        - `distance` 요청 좌표 기준 반경 3km 이내 (`lat`·`lng` 필수) / `commented` 코멘트 수 상위 30%

        `ggukPick`의 판정 기준은 홈 카드 덱과 동일하다.
        **장소분류 라벨(`labelGroup`)은 이 응답에 없다** — 라벨은 홈 카드 덱이 부여하는 값이다.

        카드형/리스트형 뷰 전환은 기획 미확정이나, 조회 계약이 아니라 클라 표현 문제로 본다.
      parameters:
        - name: roomId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [all, ggukPick, latest, distance, commented]
            default: all
        - name: category
          in: query
          required: false
          description: 지도 상단 카테고리 필터. 3종 고정
          schema:
            type: string
            enum: [all, cafe, restaurant]
            default: all
        - name: lat
          in: query
          required: false
          description: "`sort=distance`일 때 필수"
          schema:
            type: number
            format: double
        - name: lng
          in: query
          required: false
          description: "`sort=distance`일 때 필수"
          schema:
            type: number
            format: double
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK. 전체 조회 시 `pagination`은 생략한다
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Pin"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/pins/{pinId}:
    get:
      tags: [Pin]
      summary: 장소(핀) 상세 조회
      description: 장소 정보(places 컬럼 전체) + 출처 링크 + 저장한 멤버 프로필. 핀이 속한 방의 멤버만 조회할 수 있다.
      parameters:
        - $ref: "#/components/parameters/PinId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PinDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/pins/{pinId}/duplicate:
    post:
      tags: [Pin]
      summary: 다른 방에 핀 복제 ("다른 방에 공유")
      description: |
        원본 핀의 장소를 대상 방들에 복제한다. 신규 저장 경로가 아니다(신규 저장은 링크 분석 워커가 처리).
        원본 핀의 방과 **모든 대상 방**의 멤버십을 검증한다.
        **대상 방 중 하나라도 같은 장소가 이미 저장돼 있으면 `409`로 전체 거절한다.**
      parameters:
        - $ref: "#/components/parameters/PinId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roomIds]
              properties:
                roomIds:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    format: uuid
                  description: 복제 대상 방 목록
      responses:
        "200":
          description: 복제 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: 대상 방 중 하나 이상에 같은 장소가 이미 저장됨 — 전체 거절
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/pins/{pinId}/accesses:
    post:
      tags: [Pin]
      summary: 핀 접근 기록 (사용자별)
      description: |
        장소 상세 진입 시 요청 유저 기준으로 접근 행을 남긴다.
        홈 카드 덱의 묵힘(`coalesce(마지막으로 열어본 시점, 저장일)`) 계산과
        "친구들이 많이 본 곳"(클릭수) 집계의 원천 데이터를 겸한다.
        스와이프는 기록하지 않는다 — 넘긴 카드는 아직 안 본 것으로 남는다.
      parameters:
        - $ref: "#/components/parameters/PinId"
      responses:
        "200":
          description: 기록 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Firebase Authentication ID 토큰. 클라이언트가 익명 로그인(`signInAnonymously`)으로
        발급받아 보내며, 서버는 uid로 `users`를 조회해 요청 유저를 식별한다.
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PinId:
      name: pinId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: 페이지 번호 (0부터 시작)
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 페이지 크기
  responses:
...
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 권한 없음 (멤버십/방장/작성자 검증 실패)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 대상 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [errorCode, message]
      properties:
        errorCode:
          type: string
          example: FORBIDDEN
        message:
          type: string
    OkResponse:
      type: object
      description: 본문 데이터가 없는 성공 응답의 공통 포맷
      properties:
        data:
          type: object
          required: [ok]
          properties:
            ok:
              type: boolean
              example: true
    Pagination:
      type: object
      description: 목록 API 공통 페이지네이션 메타 (offset 기반)
      required: [pageSize, page, hasNext]
      properties:
        pageSize:
          type: integer
          example: 20
        page:
          type: integer
          description: 0부터 시작
          example: 0
        hasNext:
          type: boolean
    Nickname:
      type: string
      minLength: 2
      maxLength: 15
      pattern: "^[가-힣A-Za-z]+$"
      description: 한글/영문 2~15자 — 공백·숫자·특수문자 불가(안드 스펙 정합). 상한 15자는 서버(DB) 계약
    InvitationCode:
      type: string
      minLength: 6
      maxLength: 6
      pattern: "^[A-Z0-9]{6}$"
      example: "K7Q2MZ"
      description: 초대 링크 gguk.org/r/{code}의 code (대문자 영문 + 숫자 6자 고정)
    RoomName:
      type: string
      minLength: 1
      maxLength: 15
      pattern: "^[가-힣ㄱ-ㅎㅏ-ㅣA-Za-z0-9 ]+$"
      description: |
        한글·영문·숫자·공백만 허용. 앞뒤 공백을 제거한 뒤 1~15자(grapheme)이며 중복은 허용한다 (PR 리뷰 확정)
    ColorKey:
      type: string
      description: |
        색상 팔레트 키(총 13색, snake_case). 실제 색 매핑(그라데이션 등)은 클라이언트가 담당한다.
        gray는 개인방 기본값(클라이언트는 회색 기본 이미지로 렌더링).
      enum:
        - red
        - red_orange
        - orange
        - lime
        - green
        - cyan
        - violet
        - pink
        - blue
        - brown
        - light_blue
        - purple
        - gray
      example: red
    Avatar:
      type: object
      description: 프로필 아바타. 색상 키 기반으로 저장하며(실제 색 매핑은 클라이언트), 추후 필드 확장 가능
      required: [color]
      properties:
        color:
          $ref: "#/components/schemas/ColorKey"
...
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [personal, shared]
        name:
          $ref: "#/components/schemas/RoomName"
        description:
          type: string
          maxLength: 30 # grapheme 단위(클라이언트 카운터와 동일)
          nullable: true
        color:
          $ref: "#/components/schemas/ColorKey"
        ownerId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    RoomSummary:
      allOf:
        - $ref: "#/components/schemas/Room"
        - type: object
          properties:
            pinCount:
              type: integer
            memberCount:
              type: integer
            thumbnailList:
              type: array
              description: 최근 핀 최대 4개의 장소 대표 이미지 URL(최신순). 저장된 핀이 없으면 방 대표 색상 키 1개(클라이언트는 색+캐릭터로 렌더링)
              items:
                type: string
              example: ["https://img.example.com/a.jpg", "https://img.example.com/b.jpg"]
            hasPlace:
              type: boolean
              nullable: true
              description: '`?showHasPlaceId=` 지정 시에만 포함 — 해당 장소가 이 방에 이미 저장돼 있는지'
            matchedPinId:
              type: string
              format: uuid
              nullable: true
              description: '`?showHasPlaceId=` 지정 시 활성 매칭 핀 UUID, 없으면 null'
            users:
              type: array
              nullable: true
              description: '`?showUsers=true` 지정 시에만 포함 — 방 멤버 목록'
              items:
                $ref: "#/components/schemas/RoomMember"
    RoomDetail:
      allOf:
        - $ref: "#/components/schemas/Room"
        - type: object
          properties:
            pinCount:
              type: integer
            memberCount:
              type: integer
    RoomMember:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        nickname:
          $ref: "#/components/schemas/Nickname"
        avatar:
          $ref: "#/components/schemas/Avatar"
        isOwner:
          type: boolean
        joinedAt:
          type: string
          format: date-time
...
    PlaceBase:
      type: object
      description: 핀·카드 응답이 공유하는 장소 필드. images는 제외 — pins로 이동 예정(DB 변경)
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
          enum: [kakao, google]
        providerPlaceId:
          type: string
        name:
          type: string
        address:
          type: string
        city:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        lat:
          type: number
        lng:
          type: number
        category:
          type: string
          nullable: true
          description: |
            provider 원본 카테고리 문자열. 지도 상단 필터는 이 값을
            `cafe`/`restaurant`/그 외 3분류로 접어서 판정한다 (매핑 규칙 정의 예정)
        phone:
          type: string
          nullable: true
        mapUrl:
          type: string
          nullable: true
          description: 외부 지도 상세 URL (스키마 컬럼 `external_url` → `map_url` 변경 예정)
    Place:
      description: 핀 응답의 장소. 카드 응답과 달리 타임스탬프를 포함한다
      allOf:
        - $ref: "#/components/schemas/PlaceBase"
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    Pin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        place:
          $ref: "#/components/schemas/Place"
        images:
          type: array
          description: 게시물 이미지 — places.images가 pins로 이동 예정(DB 변경)
          items:
            type: string
            format: uri
        createdBy:
          $ref: "#/components/schemas/PinAuthor"
        createdAt:
          type: string
          format: date-time
    PinAuthor:
      type: object
      nullable: true
      description: 핀을 저장한 멤버의 프로필 ("누가 추가한 곳" 표시용). 탈퇴 등으로 없으면 null
      properties:
        userId:
          type: string
          format: uuid
        nickname:
          $ref: "#/components/schemas/Nickname"
        avatar:
          $ref: "#/components/schemas/Avatar"
    PinDetail:
      allOf:
        - $ref: "#/components/schemas/Pin"
        - type: object
          properties:
            sourceUrl:
              type: string
              format: uri
              nullable: true
              description: 출처 링크 (sources.original_url, 단일)
...